+++ android-unity-wrapper/unitywrapper/src/main/java/com/appsflyer/unity/AppsFlyerAndroidWrapper.java
@@ -395,4 +395,8 @@ public class AppsFlyerAndroidWrapper {
             }
         });
     }
+
+    public static void addPushNotificationDeepLinkPath(String ... path){
+        AppsFlyerLib.getInstance().addPushNotificationDeepLinkPath(path);
+    }
 }