+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Versions
 
+## v6.1.3
+
+* RD-50954 - Added Unified Deep Linking API
+* RD-54264 - Added addPushNotificationDeepLinkPath api for iOS & Android
+* RD-54266 - iOS SDK Version - 6.1.3
+* RD-54266 - Android SDK Version - 6.1.3
+
 ## v6.1.0
 
 * iOS SDK Version - 6.1.1